+++ src/main/java/com/ll/hereispaw/domain/member/member/entity/Member.java
@@ -63,7 +63,7 @@ public class Member extends BaseEntity {
 
     private String avatar;
 
-    private Integer radius;
+    private Double radius;
 
     public boolean isAdmin() {
         return "admin".startsWith(username);
